Draw Happy Diwali in Python Turtle
We will Draw Happy Diwali in Python Turtle. Turtle is a pre-installed Python module that allows users to draw drawings and shapes on a virtual canvas. The onscreen pen used for doodling is known as the turtle. In summary, the Python turtle library provides novice programmers with a fun and interactive approach to learning about Python programming.